+++ util/rand_fast.go
@@ -45,13 +45,13 @@ func (random) Uint64() uint64 {
 	return uint64(FastRand())<<32 | uint64(FastRand())
 }
 func (r random) Intn(n int) int {
-	return int(int(abs64(r.Uint64())) % n)
+	return int(abs64(r.Uint64())) % n
 }
 func (random) Int31n(n int32) int32 {
-	return int32(int32(abs32(FastRand())) & n)
+	return int32(abs32(FastRand())) & n
 }
 func (r random) Int63n(n int64) int64 {
-	return int64(int64(abs64(r.Uint64())) % n)
+	return int64(abs64(r.Uint64())) % n
 }
 func (random) Read(p []byte) (int, error) {
 	for i := range p {
